refactor(channels): make configured binding registry canonical
