+++ src/channels/plugins/configured-binding-builtins.ts
@@ -1,7 +1,7 @@
 /**
  * Configured binding built-in registration.
  *
- * Registers core configured binding consumers exactly when the registry facade needs them.
+ * Registers core configured binding consumers exactly when the registry needs them.
  */
 import { acpConfiguredBindingConsumer } from "./acp-configured-binding-consumer.js";
 import { registerConfiguredBindingConsumer } from "./configured-binding-consumers.js";
