+++ src/gateway/server-plugin-bootstrap.ts
@@ -1,7 +1,7 @@
 // Gateway plugin bootstrap helpers.
 // Applies activation config and loads the process-root plugin registry.
 import type { AmbientEnvTriggerPolicy } from "../channels/config-presence.js";
-import { primeConfiguredBindingRegistry } from "../channels/plugins/binding-registry.js";
+import { primeConfiguredBindingRegistry } from "../channels/plugins/configured-binding-registry.js";
 import { applyPluginAutoEnable } from "../config/plugin-auto-enable.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import type { ChannelPluginLoadIntent } from "../plugins/loader-types.js";
